+++ src/app/locales/en/components/archive-properties.js
@@ -1,7 +1,4 @@
 export default {
-  close: 'Close',
-  cancel: 'Cancel',
-  submit: 'Submit changes',
-  archivePropertiesHeader: 'Archive properties',
-  modifyingArchive: 'modifying archive properties',
+  discard: 'Discard changes',
+  submit: 'Save',
 };