+++ src/tests/integration/components/archive-properties-test.js
@@ -4,11 +4,11 @@ import { setupRenderingTest } from 'ember-mocha';
 import { render, find } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import {
-  getBrowsableArchiveName,
   createArchive,
   createDataset,
 } from '../../helpers/datasets-archives';
 import { run } from '@ember/runloop';
+import ArchivePropertiesViewModel from 'oneprovider-gui/utils/archive-properties-view-model';
 
 describe('Integration | Component | archive properties', function () {
   setupRenderingTest();
@@ -16,17 +16,6 @@ describe('Integration | Component | archive properties', function () {
   beforeEach(async function () {
     await run(() => createDataset(this));
     await run(() => createArchive(this));
-  });
-
-  it('renders "Archive properties" text and archive name in header', async function () {
-    const name = await getBrowsableArchiveName(this);
-
-    await renderComponent(this);
-
-    const header = find('.archive-properties-modal-header');
-    expect(header).to.exist;
-    expect(header.textContent).to.contain('Archive properties');
-    expect(header.textContent).to.contain(name);
   });
 
   it('renders "Description" field with label and textarea in edit mode', async function () {
@@ -79,25 +68,34 @@ describe('Integration | Component | archive properties', function () {
   });
 });
 
+/**
+ * @param {Mocha.Context} testCase
+ */
 async function renderComponent(testCase) {
   if (!testCase.get('spacePrivileges')) {
     testCase.set('spacePrivileges', {
       viewArchives: true,
     });
   }
+  const space = {
+    privileges: testCase.get('spacePrivileges'),
+  };
+  const viewModel = ArchivePropertiesViewModel.create({
+    ownerSource: testCase.owner,
+    space,
+    browsableArchive: testCase.get('browsableArchive'),
+  });
+  testCase.set('viewModel', viewModel);
   await render(hbs `
     {{#one-pseudo-modal id="pseudo-modal-id" as |modal|}}
-      {{archive-properties
-        browsableArchive=browsableArchive
-        modal=modal
-        spacePrivileges=spacePrivileges
-        onClose=onClose
-        onSubmit=onSubmit
-      }}
+      {{archive-properties viewModel=viewModel}}
     {{/one-pseudo-modal}}
   `);
 }
 
+/**
+ * @param {Mocha.Context} testCase
+ */
 function whenInEditMode(testCase) {
   testCase.set('spacePrivileges', {
     manageDatasets: true,